+++ photutils/utils/__init__.py
@@ -11,3 +11,4 @@ from .errors import *  # noqa: F401, F403
 from .exceptions import *  # noqa: F401, F403
 from .footprints import *  # noqa: F401, F403
 from .interpolation import *  # noqa: F401, F403
+from .wcs import *  # noqa: F401, F403